requirements.txt
requests
gradio
Pillow